refactor: move AnthropicAccumulator back to usage package
AnthropicAccumulator is usage-accumulation logic, not a wire format type.
Revert it from wire/anthropic.go to usage/accumulator.go and restore all
callers (stream handlers, usage_test.go) to use usage.NewAnthropicAccumulator.

wire/anthropic.go now contains only the two pure JSON struct types:
AnthropicMsgWire and AnthropicUsageWire.
